+++ src/WinTenBot/WinTenDev.Zizi.Models/Types/LearnData.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Label { get; set; }
         public string Message { get; set; }
-        public int FromId { get; set; }
+        public long FromId { get; set; }
         public long ChatId { get; set; }
         public string TimeStamp { get; set; }
     }
