+++ src/WinTenBot/WinTenDev.Zizi.Utils/Telegram/MemberUtil.cs
@@ -4,7 +4,7 @@ namespace WinTenDev.Zizi.Utils.Telegram
 {
     public static class MemberUtil
     {
-        public static string GetNameLink(this int userId, string name)
+        public static string GetNameLink(this long userId, string name)
         {
             return $"<a href='tg://user?id={userId}'>{name}</a>";
         }
